chore: move code snippet parsing logic into builder
