+++ src/markdown/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod code;
 pub(crate) mod elements;
 pub(crate) mod parse;
 pub(crate) mod text;
